Create Arithmetic Calculation with GUI
It is a GUI program for the basic arithmetic calculation of 2 numbers and showing the result of their addition or subtraction or multiplication or division. you may clear all using AC button and can stop the program by quit. Just a basic formation for showing how to use tkinter effectively in giving input and getting the answer!