Create GUI Marksheet
Marksheet of a Student showing the SGPA he/she got in the exams! Enter the details of the student like name, roll no, registration no and upload the grades(A or B or C or D or E or F) for the SGPA and the credits obtained in the subjects English, Hindi, Science and Mathematics! Bye-Bye!